Strip local-config and function annotations
When kustomize calls a KRM function, the input includes a set of
annotations. Many of these should be passed through to the resulting
generated object, but ’config.kubernetes.io/local-config` must be
stripped or the resulting object will be considered "local" and not
rendered as output to kustomize. Likewise, leaving the generator
configuration in `config.kubernetes.io/function` is not needed and may
be problematic if passed on to further generators.

Fixes #36

Signed-off-by: Jim Ramsay <i.am@jimramsay.com>